flairremoval/requirements.txt
certifi==2019.9.11
chardet==3.0.4
idna==2.8
loguru==0.3.2
praw==6.4.0
prawcore==1.0.1
python-dotenv==0.10.3
requests==2.22.0
six==1.12.0
update-checker==0.16
urllib3==1.25.6
websocket-client==0.56.0
